refactor(core): split session actor in different files
