+++ core/src/actors/session/messages.rs
@@ -1,0 +1,11 @@
+use actix::Message;
+
+/// Message result of unit
+pub type SessionUnitResult = ();
+
+/// Message to indicate that the session needs to send a GetPeers message through the network
+pub struct GetPeers;
+
+impl Message for GetPeers {
+    type Result = SessionUnitResult;
+}
